feat: allow connection users to expire connections
This allows connection owners to detect some application-specific
condition that determines this connection is no longer valid.  When
checked in, the pool will determine it is expired and abandon it.

In easy_stalk's case, we can more quickly move on to other servers when
it is determine that there is no available work for the given connection